Rework TMPFILE usage
Many file systems don't support TMPFILE which I thought wouldn't matter, but it turns out I'd like to use ringboard on NFS (which isn't supported). This is unfortunate because the alternatives suck:
- Use TMPFILE on /tmp. This assumes that /tmp uses a file system which supports TMPFILE (pretty reasonable I'd say), but I also discovered that some machines are configured with a very small /tmp size like 4GB.
- Use a memfd. This limits what you can store to stuff that fits in RAM. Realistically this is more than fine, but it still makes me sad.
- Use a normal file in the current directory and delete when we're done. Icky as it's user visible and also very inefficient unless we reuse the file when doing stuff in a loop.

I did some benchmarking and discovered that just creating a memfd is 10x slower than truncating it so that suggests all these tmp files we're creating should actually be reused anyway.

Thus, the strategy I went for is to convert TMPFILE uses that would fit in RAM to a memfd and otherwise a TMPFILE in the current directory with a fallback materialized hidden file.

Unfortunately we can't really reuse the files created in hot loops and sent to the server because obviously the server needs to use them. We could work around this by correlating when the server replies with a pool of open files to be reused, but that seems annoyingly complicated and I'm lazy so nah.