Move pasted items to front
The reason I didn't do this initially is because I was going to implement updating the deduplicator functionality through the broadcast API and put the MoveToFront call on the client side. Realistically, I'm not going to get around to build the broadcast API anytime soon because the current hacks I have in place work 99% of the time. So just add one more hack to implement moving to front in the watcher.