Use a slightly less crappy search cancellation architecture
In the old model, the search handler would return a token for the originator to use when cancellation was needed. This is dumb because the originator wants to cancel the search immediately when firing off a new search rather than waiting for the old search to give it back a cancellation token. In the new architecture, we invert this control and make the originator give the search a cancellation token for it to monitor. This is much nicer as we no longer need to keep track of in-flight queries. We still have head-of-line blocking problems but ah well.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>